+++ server/src/resources/transaction/transaction.controllers.js
@@ -3,70 +3,9 @@ const moment = require('moment');
 const { getMonthName } = require('../../../lib/util/date');
 const types = require('../../../lib/util/types');
 const { db } = require('../../utils/db');
+const { transform } = require('../../utils/transactionTransform');
 
 const { sequelize, Sequelize, Transaction } = db;
-
-function serializeTransactions(transactions) {
-  const serializedTransactions = {};
-
-  // serialize the shit out of this data:
-  // may be able to do this in a more sql-y way at some point?
-  transactions.forEach(item => {
-    const itemDate = new Date(item.date);
-    const monthNum = moment(itemDate).format('M');
-    const transitPassPurchasePeriod =
-      moment(itemDate)
-        .endOf('month')
-        .diff(itemDate, 'days') <= 8;
-
-    let year = moment(itemDate).format('YYYY');
-    let month = getMonthName(moment(itemDate).month());
-
-    // if a transit pass was purchased 8 days before the end of the month,
-    // move it to the next month because it was for that month that it was
-    // purchased.
-
-    // There is a bug here: amount sometimes reverts to 0.
-
-    if (item.type === types.TRANSIT_PASS_LOAD && transitPassPurchasePeriod) {
-      const addMonth = moment(itemDate, 'M').add(1, 'months');
-      if (monthNum === 12) {
-        year += 1;
-      }
-
-      month = getMonthName(moment(addMonth).month());
-    }
-
-    // god this is a mess. I apologize to my future self
-    // that learns more about sequelize.
-    if (serializedTransactions[year]) {
-      if (serializedTransactions[year][month]) {
-        if (item.type !== types.TRANSIT_PASS_LOAD) {
-          serializedTransactions[year][month].transactions.push(item);
-          serializedTransactions[year][month].amount += parseFloat(item.amount);
-        } else {
-          serializedTransactions[year][month].transitPassAmount += parseFloat(item.amount);
-        }
-      } else {
-        serializedTransactions[year][month] = {
-          transactions: item.type === types.TRANSIT_PASS_LOAD ? [] : [item],
-          amount: item.type === types.TRANSIT_PASS_LOAD ? 0 : parseFloat(item.amount),
-          transitPassAmount: item.type === types.TRANSIT_PASS_LOAD ? parseFloat(item.amount) : 0
-        };
-      }
-    } else {
-      serializedTransactions[year] = {
-        [month]: {
-          transactions: item.type === types.TRANSIT_PASS_LOAD ? [] : [item],
-          amount: item.type === types.TRANSIT_PASS_LOAD ? 0 : parseFloat(item.amount),
-          transitPassAmount: item.type === types.TRANSIT_PASS_LOAD ? parseFloat(item.amount) : 0
-        }
-      };
-    }
-  });
-
-  return serializedTransactions;
-}
 
 const monthly = async (req, res, next) => {
   try {
@@ -74,6 +13,7 @@ const monthly = async (req, res, next) => {
     const searchDateMin = `${parseInt(month, 10) === 1 ? parseInt(year, 10) - 1 : year}-${
       parseInt(month, 10) === 1 ? '12' : parseInt(month, 10) - 1
     }-01`;
+
     const searchDateMax =
       month === '12' ? `${parseInt(year, 10) + 1}-01-01` : `${year}-${parseInt(month, 10) + 1}-01`;
     const transactions = await Transaction.findAll({
@@ -95,15 +35,15 @@ const monthly = async (req, res, next) => {
       order: sequelize.literal('date DESC')
     });
 
-    const serializedTransactions = serializeTransactions(transactions);
+    const transformedData = transform(transactions);
     const totalAmount =
-      serializedTransactions[year][getMonthName(month - 1)].amount +
-      serializedTransactions[year][getMonthName(month - 1)].transitPassAmount;
-    const totalTrips = serializedTransactions[year][getMonthName(month - 1)].transactions.length;
+      transformedData[year][getMonthName(month - 1)].amount +
+      transformedData[year][getMonthName(month - 1)].transitPassAmount;
+    const totalTrips = transformedData[year][getMonthName(month - 1)].transactions.length;
     res.json({
       status: 'success',
       data: {
-        transactions: serializedTransactions[year][getMonthName(month - 1)].transactions,
+        transactions: transformedData[year][getMonthName(month - 1)].transactions,
         totalTrips,
         totalAmount
       }
@@ -135,9 +75,9 @@ const all = async (req, res, next) => {
     });
 
     console.log('raw transactions:', transactions);
-    const serializedTransactions = serializeTransactions(transactions);
+    const transformedData = transform(transactions);
 
-    res.json({ status: 'success', data: serializedTransactions });
+    res.json({ status: 'success', data: transformedData });
   } catch (err) {
     next(err);
   }
@@ -176,9 +116,9 @@ const ytdData = async (req, res, next) => {
       order: sequelize.literal('date ASC')
     });
 
-    const serializedTransactions = serializeTransactions(transactions);
-    console.log(serializedTransactions);
-    res.json({ status: 'success', data: serializedTransactions });
+    const transformedData = transform(transactions);
+    console.log(transformedData);
+    res.json({ status: 'success', data: transformedData });
   } catch (error) {
     next(error);
   }
